[bug fix] parse out function argument from json
function call args are a serialized json object with named arguments.
Need to parse it out to pass the actual value into the function.

Tested by printing out the result of the function call before and after.

Before, since the argument is wrong, it gets no matches and you get empty list.
Now you get a book record as expected.