add py version of app, include text gen step
Repurpose the aiconfig prompt to be a RAG-like step 1, add second prompt for step 2.
Implement the python app to run the 2 steps.


Tests
- (regression test, run ts app)
- py:

`python app.py "how widely-read is 'To Kill a Mockingbird'?"`

out:

```
...
Response:

According to the provided data, "To Kill a Mockingbird" has sold over 18 million copies and has been translated into forty languages. It is considered a masterpiece of American literature and is said to have universal appeal. From this information, it can be inferred that the book is widely-read and has a significant readership.

```